Add 5 Files Related To The Use Of Async and Await
new file:
    - async-await-asynchronous-function-example-1-a.js
    - async-await-asynchronous-function-example-1-b.js
    - async-await-asynchronous-function-example-1-c.js
    - content/result-mind-grenade-2.txt
    - content/result-mind-grenade.txt

Each async... file improves upon the code of the previous async... file, demonstrating that it is better to use async and await instead of a bunch of nested callbacks, promises, and such.

The other two files are the result of running the code the async...b and async...c files.