Add File 7-params-query.js
This file contains basic examples of parameters and query parameters to demonstrate how to use them and return data based on their input, values. There are also notes and comments concerning these things within this file too.

Near the end is a demonstration of how to search, sort the data and limit the output with query parameters.